Move to Elasticsearch Beyonder project
The goal is to create two projects from this one so one can use the new project in another context than Spring.

The new elasticsearch-beyonder project lives now at https://github.com/dadoonet/elasticsearch-beyonder.

The goal is to simplify more and more the Spring project as soon as new features will be ported to the Beyonder project.